Add Android spawn() “uid” option to specify user ID

+++ src/linux/agent/system-server.js
@@ -1,6 +1,7 @@
-var ApplicationInfo, ComponentName, Intent, RunningAppProcessInfo, RunningTaskInfo, GET_META_DATA, GET_ACTIVITIES;
+var ApplicationInfo, ComponentName, ContextWrapper, Intent, RunningAppProcessInfo, RunningTaskInfo, UserHandle, GET_META_DATA, GET_ACTIVITIES;
 var context, packageManager, activityManager;
 
+var multiUserSupported;
 var pendingLaunches = {};
 
 function init() {
@@ -8,14 +9,18 @@ function init() {
   var ActivityThread = Java.use('android.app.ActivityThread');
   ApplicationInfo = Java.use('android.content.pm.ApplicationInfo');
   ComponentName = Java.use('android.content.ComponentName');
+  ContextWrapper = Java.use('android.content.ContextWrapper');
   Intent = Java.use('android.content.Intent');
   var Context = Java.use('android.content.Context');
   var PackageManager = Java.use('android.content.pm.PackageManager');
   RunningAppProcessInfo = Java.use('android.app.ActivityManager$RunningAppProcessInfo');
   RunningTaskInfo = Java.use('android.app.ActivityManager$RunningTaskInfo');
+  UserHandle = Java.use('android.os.UserHandle');
   var ACTIVITY_SERVICE = Context.ACTIVITY_SERVICE.value;
   GET_META_DATA = PackageManager.GET_META_DATA.value;
   GET_ACTIVITIES = PackageManager.GET_ACTIVITIES.value;
+
+  multiUserSupported = 'getApplicationInfoAsUser' in PackageManager;
 
   context = ActivityThread.currentApplication();
 
@@ -75,19 +80,37 @@ rpc.exports = {
       return result;
     });
   },
-  getProcessName: function (pkg) {
+  getProcessName: function (pkg, uid) {
+    checkUidOptionSupported(uid);
+
     return performOnJavaVM(function () {
       try {
-        return packageManager.getApplicationInfo(pkg, GET_META_DATA).processName.value;
+        return getAppMetaData(pkg, uid).processName.value;
       } catch (e) {
-        throw new Error("Unable to find application with identifier '" + pkg + "'");
+        throw new Error("Unable to find application with identifier '" + pkg + "'" +
+            ((uid !== 0) ? ' belonging to uid ' + uid : ''));
       }
     });
   },
-  startActivity: function (pkg, activity) {
+  startActivity: function (pkg, activity, uid) {
+    checkUidOptionSupported(uid);
+
     return performOnJavaVM(function () {
+      var user, ctx, pm;
+      if (uid !== 0) {
+        user = UserHandle.of(uid);
+        ctx = context.createPackageContextAsUser(pkg, GET_META_DATA, user);
+        pm = ctx.getPackageManager();
+      } else {
+        user = null;
+        ctx = context;
+        pm = packageManager;
+      }
+
       var appInstalled = false;
-      var apps = packageManager.getInstalledApplications(GET_META_DATA);
+      var apps = (uid !== 0)
+          ? pm.getInstalledApplicationsAsUser(GET_META_DATA, uid)
+          : pm.getInstalledApplications(GET_META_DATA);
       var numApps = apps.size();
       for (var i = 0; i !== numApps; i++) {
         var appInfo = Java.cast(apps.get(i), ApplicationInfo);
@@ -99,9 +122,9 @@ rpc.exports = {
       if (!appInstalled)
         throw new Error("Unable to find application with identifier '" + pkg + "'");
 
-      var intent = packageManager.getLaunchIntentForPackage(pkg);
-      if (intent === null && 'getLeanbackLaunchIntentForPackage' in packageManager)
-        intent = packageManager.getLeanbackLaunchIntentForPackage(pkg);
+      var intent = pm.getLaunchIntentForPackage(pkg);
+      if (intent === null && 'getLeanbackLaunchIntentForPackage' in pm)
+        intent = pm.getLeanbackLaunchIntentForPackage(pkg);
       if (intent === null && activity === null)
         throw new Error('Unable to find a front-door activity');
 
@@ -111,7 +134,9 @@ rpc.exports = {
       }
 
       if (activity !== null) {
-        var pkgInfo = packageManager.getPackageInfo(pkg, GET_ACTIVITIES);
+        var pkgInfo = (uid !== 0)
+            ? pm.getPackageInfoAsUser(pkg, GET_ACTIVITIES, uid)
+            : pm.getPackageInfo(pkg, GET_ACTIVITIES);
         var activities = pkgInfo.activities.value.map(function (activityInfo) {
           return activityInfo.name.value;
         });
@@ -121,25 +146,38 @@ rpc.exports = {
         intent.setClassName(pkg, activity);
       }
 
-      performLaunchOperation(pkg, function () {
-        context.startActivity(intent);
+      performLaunchOperation(pkg, uid, function () {
+        if (user !== null)
+          ContextWrapper.$new(ctx).startActivityAsUser(intent, user);
+        else
+          ctx.startActivity(intent);
       });
     });
   },
-  sendBroadcast: function (pkg, receiver, action) {
+  sendBroadcast: function (pkg, receiver, action, uid) {
+    checkUidOptionSupported(uid);
+
     return performOnJavaVM(function () {
       var intent = Intent.$new();
       intent.setComponent(ComponentName.$new(pkg, receiver));
       intent.setAction(action);
 
-      performLaunchOperation(pkg, function () {
-        context.sendBroadcast(intent);
+      performLaunchOperation(pkg, uid, function () {
+        if (uid !== 0)
+          ContextWrapper.$new(context).sendBroadcastAsUser(intent, UserHandle.of(uid));
+        else
+          context.sendBroadcast(intent);
       });
     });
   },
-  stopPackage: function (pkg) {
+  stopPackage: function (pkg, uid) {
+    checkUidOptionSupported(uid);
+
     return performOnJavaVM(function () {
-      activityManager.forceStopPackage(pkg);
+      if (uid !== 0)
+        activityManager.forceStopPackageAsUser(pkg, uid);
+      else
+        activityManager.forceStopPackage(pkg);
     });
   },
   tryStopPackageByPid: function (pid) {
@@ -167,7 +205,7 @@ rpc.exports = {
       var runningTaskInfos = activityManager.getRunningTasks(1);
       if (runningTaskInfos !== null && runningTaskInfos.size() > 0) {
         var runningTaskInfo = Java.cast(runningTaskInfos.get(0), RunningTaskInfo);
-        if (typeof runningTaskInfo.topActivity !== 'undefined') {
+        if (runningTaskInfo.topActivity !== undefined) {
           var topActivity = runningTaskInfo.topActivity.value;
           var app = packageManager.getApplicationInfo(topActivity.getPackageName(), GET_META_DATA);
           var pkg = app.packageName.value;
@@ -193,6 +231,17 @@ rpc.exports = {
     });
   }
 };
+
+function getAppMetaData(pkg, uid) {
+  return (uid !== 0)
+      ? packageManager.getApplicationInfoAsUser(pkg, GET_META_DATA, uid)
+      : packageManager.getApplicationInfo(pkg, GET_META_DATA);
+}
+
+function checkUidOptionSupported(uid) {
+  if (uid !== 0 && !multiUserSupported)
+    throw new Error('The “uid” option is not supported on the current Android OS version');
+}
 
 function installLaunchTimeoutRemovalInstrumentation() {
   var Handler = Java.use('android.os.Handler');
@@ -224,8 +273,8 @@ function installLaunchTimeoutRemovalInstrumentation() {
   };
 }
 
-function performLaunchOperation(pkg, operation) {
-  var processName = packageManager.getApplicationInfo(pkg, GET_META_DATA).processName.value;
+function performLaunchOperation(pkg, uid, operation) {
+  var processName = getAppMetaData(pkg, uid).processName.value;
 
   tryFinishLaunch(processName);
